requirements.txt
anyio==4.6.0
attrs==24.2.0
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
mysql-connector-python==9.0.0
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pandas-ta==0.3.14b0
pillow==10.4.0
plotly==5.24.1
prompt_toolkit==3.0.48
pycparser==2.22
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
schedule==1.2.2
schwab-py==0.0.0a25
scipy==1.15.2
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.0.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
zipp==3.20.2
psycopg2-binary==2.9.10
mplcursors==0.6.0
